requirements.txt
flask
python-dotenv
pymongo[srv]
gunicorn
flask-wtf
flask-security
email_validator
passlib